requirements.txt
streamlit==1.28.1
pandas==2.1.3
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.11.0
pptx==0.6.21
googlemaps==4.10.0
